# static/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Idling.app API",
    "version": "0.338.7",
    "description": "Complete API documentation for Idling.app - a modern social platform",
    "contact": {
      "name": "Idling.app Development Team",
      "url": "https://github.com/Underwood-Inc/idling.app__UI"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server (Next.js)"
    },
    {
      "url": "https://idling.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "NextAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie"
      },
      "AdminRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Role",
        "description": "Admin role verification"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "User name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email"
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "User avatar URL"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation date"
          },
          "profile_public": {
            "type": "boolean",
            "description": "Whether profile is public"
          },
          "bio": {
            "type": "string",
            "description": "User bio"
          },
          "location": {
            "type": "string",
            "description": "User location"
          },
          "last_login": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last login date"
          },
          "login_count": {
            "type": "number",
            "description": "Number of logins"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether user is active"
          },
          "admin_notes": {
            "type": "string",
            "nullable": true,
            "description": "Admin notes about user"
          },
          "provider_name": {
            "type": "string",
            "nullable": true,
            "description": "Auth provider name"
          },
          "provider_email": {
            "type": "string",
            "nullable": true,
            "description": "Auth provider email"
          },
          "provider_verified": {
            "type": "boolean",
            "description": "Whether provider email is verified"
          },
          "provider_last_used": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last provider usage"
          },
          "role_count": {
            "type": "number",
            "description": "Number of roles assigned"
          },
          "role_names": {
            "type": "string",
            "nullable": true,
            "description": "Comma-separated role names"
          },
          "has_admin": {
            "type": "boolean",
            "description": "Whether user has admin role"
          },
          "has_moderator": {
            "type": "boolean",
            "description": "Whether user has moderator role"
          },
          "subscription_count": {
            "type": "number",
            "description": "Number of subscriptions"
          },
          "active_subscriptions": {
            "type": "number",
            "description": "Number of active subscriptions"
          },
          "subscription_names": {
            "type": "string",
            "nullable": true,
            "description": "Comma-separated subscription names"
          },
          "has_active_subscription": {
            "type": "boolean",
            "description": "Whether user has active subscription"
          },
          "is_timed_out": {
            "type": "boolean",
            "description": "Whether user is currently timed out"
          },
          "timeout_count": {
            "type": "number",
            "description": "Number of timeouts"
          },
          "active_timeout_reason": {
            "type": "string",
            "nullable": true,
            "description": "Reason for current timeout"
          },
          "timeout_expires": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When current timeout expires"
          },
          "active_timeout_id": {
            "type": "number",
            "nullable": true,
            "description": "ID of active timeout"
          }
        }
      }
    }
  },
...
  "paths": {
    "/api/admin/users": {
      "get": {
        "summary": "Get users for admin management",
        "description": "Retrieve a paginated list of users with comprehensive information for admin management",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "NextAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Number of users per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term to filter users by name, email, or provider"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminUser"
                      }
                    },
                    "currentPage": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "hasNext": {
                          "type": "boolean"
                        },
                        "hasPrev": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions (admin required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}